requirements.txt
flask
transformers
torch
pytesseract
pillow
nltk
opencv-python
numpy